feat: SQLite storage backend for trades, flips, and recipe flips
Adds an optional SQLite persistence backend alongside the JSON files,
with a one-time migration from JSON, live backend switching without a
restart, and maintenance actions (delete / regenerate database).

Storage core:
- SqliteStorage with WAL journaling, per-account data isolation, and
  schema versioning via PRAGMA user_version (v6; the single v5->v6 step
  only exists to upgrade development databases, the backend never shipped)
- trades record terminal offer events only; partial fills are never
  double-counted
- trades always store original quantities: recipe consumption is
  recorded in consumed_trade and subtracted at read time
  (reconcileFlipsForItem, recipe component loaders)
- INSERT OR IGNORE results checked via update count instead of stale
  getGeneratedKeys; writes run on a background executor; failures
  surface as IllegalStateException instead of being swallowed
- complete delete paths: deleteAccountData, deleteOffersSince,
  deleteTradesByUuid, deleteRecipeFlip, deleteRecipeFlipsSince
- both GE-limit counters (window + complete-offer base) persist, so
  limit counts survive restarts (buy 100, restart, buy 10 -> 110)

Every mutation reaches the active backend: live trades, GE-history
imports, favorite toggles and quick-search codes, and recipe flip
deletions all round-trip through SQLite when it is enabled.

Migration and switching:
- atomic per-account transactions under the storage monitor; idempotent
  and self-healing (interrupted migrations retry on next startup)
- in-memory data is flushed before any backend switch; JSON files are
  preserved and the toggle is reversible
- switching back to SQLite wipes each account and re-imports from the
  authoritative JSON, reconciling deletions made in JSON mode

Stats parity:
- aggregate stats and item summaries delegate to the same interval
  computation the JSON backend uses, so totals match exactly; recipe
  group summaries compute from interval-filtered flips
- compatible with both the current API (latest.release, 1.12.39) and
  the upcoming max-cash API (64-bit prices) via saturating casts

Tests: real-data migration parity/integrity/retry tests against actual
account files (gitignored fixtures, auto-skip when absent),
migration/live-trade interaction, flip reconciliation, schema upgrade,
and regression tests for every round-trip above.

+++ src/main/java/com/flippingutilities/ui/statistics/recipes/customrecipes/SuggestionData.java
@@ -5,10 +5,11 @@ import net.runelite.client.util.AsyncBufferedImage;
 class SuggestionData {
     final int itemId;
     final String name;
-    final int price;
+    // 64-bit since the max cash update
+    final long price;
     final AsyncBufferedImage image;
 
-    SuggestionData(int itemId, String name, int price, AsyncBufferedImage image) {
+    SuggestionData(int itemId, String name, long price, AsyncBufferedImage image) {
         this.itemId = itemId;
         this.name = name;
         this.price = price;
